requirements.txt
pandas
python-dotenv
supabase
gspread
requests